Refactored Pages
