+++ src/Components/Context/CategoryData.ts
@@ -1,0 +1,5 @@
+export interface CategoryData {
+    id: string
+    name: string
+    subcategories: string[]
+ }
