+++ src/Components/Context/ClassroomRecord.ts
@@ -1,0 +1,5 @@
+export interface ClassroomRecord {
+    name: string
+    id: number
+    tickets: number[]
+ }
