+++ src/Components/Context/TicketRecord.ts
@@ -1,0 +1,13 @@
+export interface TicketRecord {
+    title: string
+    categoryId: string
+    subcategoryIndex: number
+    priority: number
+    intermediaries: string
+    description: string
+    status: number
+    createdAt: string
+    id: number
+    closedAt: string
+    classroomId: number
+ }
